requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Agent Orchestration
langgraph==0.0.26

# Computer Vision
opencv-python-headless==4.8.1.78
pillow==10.1.0
ultralytics==8.2.0
numpy==1.26.4

# OCR & PDF
pytesseract==0.3.10
pdf2image==1.16.3
pymupdf==1.23.8
PyPDF2==3.0.1

# ML & Embeddings
sentence-transformers==2.2.2
torch==2.2.0
scikit-learn==1.3.2
scikit-image==0.22.0

# Vector Database
qdrant-client==1.6.4

# Optional LLM (hooks included)
openai==1.12.0
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
python-json-logger==2.0.7
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1